Refactored the ingestion service API and split into events and sensor endpoints .py files
