+++ ingestion_service/app/api/v1/api.py
@@ -1,6 +1,8 @@
 from fastapi import APIRouter
-from app.api.v1.endpoints import endpoints
+from app.api.v1.endpoints import sensors
+from app.api.v1.endpoints import events
 
 
 api_router = APIRouter()
-api_router.include_router(endpoints.router, prefix="/events", tags=["Events"])
+api_router.include_router(events.router, prefix="/events", tags=["Events"])
+api_router.include_router(sensors.router, prefix="/sensors", tags=["Sensors"])
